Revert "drag to insert CSV (wip)"
This reverts commit ed210c4ca11d5fcb58173e0e4a36e57693eb2129.
