+++ src/grid/actions/clipboard/clipboard.ts
@@ -7,7 +7,6 @@ import { CellAndFormat } from '../../sheet/GridSparse';
 import { Rectangle } from 'pixi.js';
 import { clearFormattingAction } from '../clearFormattingAction';
 import { clearBordersAction } from '../clearBordersAction';
-import { InsertTabSeparatedText } from '../insertData/insertTabSeparatedText';
 
 const CLIPBOARD_FORMAT_VERSION = 'quadratic/clipboard/json/1.1';
 
@@ -132,12 +131,55 @@ const pasteFromText = async (sheet_controller: SheetController, pasteToCell: Coo
   try {
     // attempt to read text from clipboard
     const clipboard_text = await navigator.clipboard.readText();
+    let cell_x: number = pasteToCell.x;
+    let cell_y: number = pasteToCell.y;
 
-    await InsertTabSeparatedText({
-      sheetController: sheet_controller,
-      text: clipboard_text,
-      insertAtCellLocation: pasteToCell,
+    // build api payload
+    let cells_to_write: Cell[] = [];
+    let cells_to_delete: Coordinate[] = [];
+
+    let str_rows: string[] = clipboard_text.split('\n');
+
+    // for each copied row
+    str_rows.forEach((str_row) => {
+      let str_cells: string[] = str_row.split('\t');
+
+      // for each copied cell
+      str_cells.forEach((str_cell) => {
+        // update or clear cell
+        if (str_cell !== '') {
+          cells_to_write.push({
+            x: cell_x,
+            y: cell_y,
+            type: 'TEXT',
+            value: str_cell,
+            last_modified: new Date().toISOString(),
+          });
+        } else {
+          cells_to_delete.push({
+            x: cell_x,
+            y: cell_y,
+          });
+        }
+
+        // move to next cell
+        cell_x += 1;
+      });
+
+      // move to next row and return
+      cell_y += 1;
+      cell_x = pasteToCell.x;
     });
+
+    // TODO ALSO BE ABLE TO PASS CELLS TO DELETE TO updatecellandcells
+
+    // bulk update and delete cells
+    await updateCellAndDCells({
+      starting_cells: cells_to_write,
+      sheetController: sheet_controller,
+    });
+
+    // cells_to_delete
 
     return true; // unsuccessful
   } catch {
