+++ src/gridGL/pixiApp/PixiApp.ts
@@ -19,7 +19,6 @@ import { SheetController } from '../../grid/controller/sheetController';
 import { HEADING_SIZE } from '../../constants/gridConstants';
 import { editorInteractionStateDefault } from '../../atoms/editorInteractionStateAtom';
 import { gridInteractionStateDefault } from '../../atoms/gridInteractionStateAtom';
-// import { FileUploadIndicator } from '../interaction/fileUpload/FileUploadIndicator';
 
 export class PixiApp {
   private parent?: HTMLDivElement;
@@ -39,7 +38,6 @@ export class PixiApp {
   quadrants: Quadrants;
 
   input: Pointer;
-  // fileUploadIndicator: FileUploadIndicator;
   viewportContents: Container;
   settings: PixiAppSettings;
   renderer: Renderer;
@@ -109,7 +107,6 @@ export class PixiApp {
 
     this.input = new Pointer(this);
     this.update = new Update(this);
-    // this.fileUploadIndicator = new FileUploadIndicator(this);
 
     if (debugAlwaysShowCache) this.showCache();
