+++ src/ui/QuadraticUI.tsx
@@ -12,7 +12,6 @@ import { useEffect, useState } from 'react';
 import { PixiApp } from '../gridGL/pixiApp/PixiApp';
 import { SheetController } from '../grid/controller/sheetController';
 import CellTypeMenu from './menus/CellTypeMenu';
-import { FileUploadWrapper } from './components/FileUploadWrapper';
 
 interface Props {
   sheetController: SheetController;
@@ -54,9 +53,7 @@ export default function QuadraticUI(props: Props) {
           position: 'relative',
         }}
       >
-        <FileUploadWrapper sheetController={sheetController} app={app}>
-          <QuadraticGrid sheetController={sheetController} app={app} />
-        </FileUploadWrapper>
+        <QuadraticGrid sheetController={sheetController} app={app} />
         <CodeEditor editorInteractionState={editorInteractionState} sheet_controller={sheetController} />
       </div>
